requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
aiohttp==3.9.1
pillow==10.1.0
requests==2.31.0
